requirements.txt
numpy>=1.24.3
Pillow>=9.5.0
customtkinter>=5.2.0